requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
pandas==2.1.0
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==20.1.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
